refactor: 온보딩 모달 배경 컴포넌트 분리
